requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
